create and define each module
